Factor out move container for future optimization

+++ src/Square.h
@@ -15,7 +15,7 @@ class Square {
 public:
   Square();
   virtual ~Square();
-  static void print_square(int square);
+  static void print_square(uint8_t square);
   static string mailbox_index_to_square(int square);
 
 };
